chore - move MainHeader
